Correções de funcionalidades que envolvem cache

+++ src/main/java/com/strongmemoryapi/service/word/WordService.java
@@ -20,8 +20,8 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 
 @Service
 public class WordService {
@@ -32,15 +32,16 @@ public class WordService {
     @Autowired
     private DifficultyService difficultyService;
 
-    @CacheEvict(value = "wordIdsByDifficulty", key = "#difficultyId")
+    @Autowired
+    private WordCacheService cacheService;
+
+    @CacheEvict(value = "wordIdsByDifficulty", allEntries = true)
     public WordResponse register(WordRegistrationRequest request){
         if(wordRepository.existsByWord(request.word())){
             throw new ResourceAlreadyExistsException("Palavra já cadastrada.");
         }
 
         DifficultyEntity difficulty = difficultyService.getByDifficultyName(request.difficulty());
-        Byte difficultyId = difficulty.getId();
-
         WordEntity word = new WordEntity();
         word.setWord(request.word());
         word.setDifficulty(difficulty);
@@ -48,22 +49,20 @@ public class WordService {
         return this.parseToWordResponse(wordRepository.save(word));
     }
 
-    @CacheEvict(value = "wordIdsByDifficulty", key = "#difficultyId")
+    @CacheEvict(value = "wordIdsByDifficulty", allEntries = true)
     public void update(Long id, WordUpdateRequest request){
         WordEntity word = wordRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Palavra não encontrada."));
 
-        Byte difficultyId = word.getDifficulty().getId();
-
         word.setWord(request.word());
-
         wordRepository.save(word);
     }
 
-    @CacheEvict(value = "wordIdsByDifficulty", key = "#difficultyId")
+    @CacheEvict(value = "wordIdsByDifficulty", allEntries = true)
     public void delete(Long id){
-        Byte difficultyId = wordRepository.findDifficultyIdById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Palavra não encontrada."));
+        if(wordRepository.existsById(id)){
+           throw new ResourceNotFoundException("Palavra não encontrada.");
+        }
         wordRepository.deleteById(id);
     }
 
@@ -74,23 +73,17 @@ public class WordService {
 
         DifficultyEntity currentDifficulty = difficultyService.getByDifficultyName(difficulty);
 
-        Long totalWords = wordRepository.countByDifficulty(currentDifficulty);
-        if(totalWords < quantityWords){
+        List<Long> wordsIds = new ArrayList<>(cacheService.findIdsByDifficulty(currentDifficulty));
+
+        if(wordsIds.size() < quantityWords){
             throw new InsufficientWordsException();
         }
 
-        List<Long> randomIds = new ArrayList<>();
-        List<Long> wordsIds = findIdsByDifficulty(currentDifficulty);
-
-        Random r = new Random();
-        for(int i = 0; i < quantityWords; i++){
-            int randomPos = r.nextInt(0, wordsIds.size() - 1);
-            Long randomId = wordsIds.remove(randomPos);
-            randomIds.add(randomId);
-        }
+        Collections.shuffle(wordsIds);
+        List<Long> selectedIds = wordsIds.subList(0, quantityWords);
 
         return wordRepository
-                .findByIdIn(randomIds)
+                .findByIdIn(selectedIds)
                 .stream()
                 .map(this::parseToWordResponse)
                 .toList();
@@ -106,14 +99,8 @@ public class WordService {
 
         Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
         return wordRepository
-                    .findAllByDifficulty(pageable, currentDifficulty)
-                    .map(this::parseToWordResponse);
-    }
-
-    @Cacheable(value = "wordIdsByDifficulty", key = "#difficultyId")
-    public List<Long> findIdsByDifficulty(DifficultyEntity difficulty) {
-        Byte difficultyId = difficulty.getId();
-        return wordRepository.findIdsByDifficulty(difficulty);
+                .findAllByDifficulty(pageable, currentDifficulty)
+                .map(this::parseToWordResponse);
     }
 
     private WordResponse parseToWordResponse(WordEntity word){
